requirements.txt
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
psycopg[binary]==3.3.2
python-dotenv==1.2.1
reportlab==4.4.10
whitenoise==6.9.0
